moved comments to rant because it was unnecessary to separate them
