--- app/src/main/js/actions/comments.js
@@ -1,29 +1,0 @@
-import {
-  COMMENT_POST,
-} from '../consts/comments';
-
-export function updateCommentPost(text) {
-  return (dispatch) => {
-    dispatch({
-      type: COMMENT_POST.UPDATE,
-      payload: text,
-    });
-  };
-}
-
-export function clearCommentPost() {
-  return (dispatch) => {
-    dispatch({
-      type: COMMENT_POST.CLEAR,
-    });
-  };
-}
-
-export function addUserCommentPost(username) {
-  return (dispatch) => {
-    dispatch({
-      type: COMMENT_POST.ADD_USER,
-      user: username,
-    });
-  };
-}
