+++ app/src/main/js/actions/rant.js
@@ -2,6 +2,7 @@ import rantscript from 'rantscript';
 import {
   FETCH_RANT,
   CLOSE_RANT,
+  COMMENT_POST,
 } from '../consts/rants';
 import { STATE } from '../consts/state';
 
@@ -30,6 +31,32 @@ export function closeRant() {
   return (dispatch) => {
     dispatch({
       type: CLOSE_RANT,
+    });
+  };
+}
+
+export function updateCommentPost(text) {
+  return (dispatch) => {
+    dispatch({
+      type: COMMENT_POST.UPDATE,
+      payload: text,
+    });
+  };
+}
+
+export function clearCommentPost() {
+  return (dispatch) => {
+    dispatch({
+      type: COMMENT_POST.CLEAR,
+    });
+  };
+}
+
+export function addUserCommentPost(username) {
+  return (dispatch) => {
+    dispatch({
+      type: COMMENT_POST.ADD_USER,
+      user: username,
     });
   };
 }
