--- app/src/main/js/consts/comments.js
@@ -1,5 +1,0 @@
-export const COMMENT_POST = {
-  UPDATE: 'COMMENT_POST_UPDATE',
-  ADD_USER: 'COMMENT_POST_ADD_USER',
-  CLEAR: 'COMMENT_POST_CLEAR',
-}
