+++ app/src/main/js/consts/rants.js
@@ -2,3 +2,8 @@ export const FETCH_RANTS = 'FETCH_RANTS';
 export const RESET_PAGE = 'RESET_PAGE';
 export const FETCH_RANT = 'FETCH_RANT';
 export const CLOSE_RANT = 'CLOSE_RANT';
+export const COMMENT_POST = {
+  UPDATE: 'COMMENT_POST_UPDATE',
+  ADD_USER: 'COMMENT_POST_ADD_USER',
+  CLEAR: 'COMMENT_POST_CLEAR',
+};
