+++ app/src/main/js/reducers/index.js
@@ -2,13 +2,11 @@ import { combineReducers } from 'redux';
 import { theme } from './settings';
 import { rants } from './rants';
 import { rant } from './rant';
-import { comments } from './comments';
 
 const rootReducer = combineReducers({
   theme,
   rants,
   rant,
-  comments,
 });
 
 export default rootReducer;
